+++ src/main/kotlin/com/springboot/riot/data/summoner/dto/MatchTeamBansDto.kt
@@ -1,7 +1,7 @@
 package com.springboot.riot.data.summoner.dto
 
 data class MatchTeamBansDto(
-    var gameId: Int = 0,
+    var gameId: Long = 0L,
     var teamId: Int = 0,
     val championId: String? = null,
     val pickTurn: String? = null,
